Update slider index.html
Tiny change to the slider example usage. TextBox takes str, and does not auto-convert the int from the slider to a str. This makes the example runnable.

In addition, I've widened the textbox, so it shows both digits of the slider's value. I am unsure if this is a platform/device specific behavior, but on my machine only the second digit of the value was shown, making the example somewhat confusing.